+++ common/src/main/java/com/faboslav/friendsandfoes/entity/IllusionerEntityAccess.java
@@ -15,8 +15,6 @@ public interface IllusionerEntityAccess
 
 	void tickMovement();
 
-	void initDataTracker();
-
 	void writeCustomDataToNbt(NbtCompound nbt);
 
 	void readCustomDataFromNbt(NbtCompound nbt);